Remove AbstractHandlerRemove MappersRemove Custom ModelsRename MessageHandlers to Handlers
