+++ lib/handlers/application_handler.rb
@@ -2,11 +2,23 @@
 
 require 'types/message_entities/bot_command'
 
-module MessageHandlers
+module Handlers
   class ApplicationHandler
+    include Import['telegram.bot']
+
+    def initialize(**params)
+      @current_message = params.delete(:message)
+
+      super(params)
+    end
+
     def call(message)
       handler_for(message).new(message: message).call
     end
+
+    private
+
+    attr_reader :current_message
 
     def handler_for(message)
       case message
