--- lib/mappers/user_mapper.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-require 'models/user'
-
-class UserMapper < ROM::Mapper
-  relation :users
-  register_as :user
-  model User
-end
