--- spec/lib/message_handlers_spec.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-require 'spec_helper'
-
-RSpec.describe MessageHandlers do
-  describe '.handle' do
-  end
-end
