--- spec/message_handlers/command_handlers/start_handler_spec.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-require 'spec_helper'
-
-RSpec.describe MessageHandlers::CommandHandlers::StartHandler do
-end
